Respawn point added

+++ FPS-Test/Assets/Scripts/Gun.cs
@@ -2,7 +2,7 @@
 
 public class Gun : MonoBehaviour {
 
-    public float damage = 10.0f;
+    public int damage = 10;
     public float range = 100.0f;
     public Camera fpsCam;
     public ParticleSystem MuzzleFlash;
